refactor: remove trips v1 code
