+++ src/api/index.ts
@@ -1,4 +1,3 @@
 export {CancelToken, isCancel, client} from './client';
 export {autocomplete, reverse} from './geocoder';
-export {search} from './trips';
 export type {SubscriptionState} from './use-subscription';
